Move preloaded statblock initialization to Libraries
